backend/requirements.txt
fastapi
uvicorn
celery
redis
sqlalchemy
psycopg2-binary
python-multipart
pydantic-settings
alembic
python-dotenv
